:rocket: update: change to underscored in db model

+++ src/controllers/vote.controller.ts
@@ -10,12 +10,14 @@ const createVote = async (req: RequestWithUserId, res: Response): Promise<Respon
   try {
     const post = await Post.findOne({ where: { id: postId } })
     if (post) {
-      const existVote = await Vote.findOne({ where: { postId, userId } })
+      const existVote = await Vote.findOne({
+        where: { post_id: postId, user_id: userId },
+      })
       if (dir === 1) {
         if (existVote) {
           return res.status(409).json({ message: 'Vote already exits' })
         }
-        await Vote.create({ postId, userId })
+        await Vote.create({ post_id: postId, user_id: userId })
         return res.status(201).json({
           message: 'Succefully added vote',
         })
@@ -23,7 +25,7 @@ const createVote = async (req: RequestWithUserId, res: Response): Promise<Respon
         if (!existVote) {
           return res.status(404).json({ message: 'Vote not exits' })
         }
-        await Vote.destroy({ where: { postId, userId } })
+        await Vote.destroy({ where: { post_id: postId, user_id: userId } })
         return res.status(201).json({
           message: 'Succefully deleted vote',
         })
